+++ lib/source2epub.rb
@@ -1,5 +1,20 @@
-require_relative "source2epub/version"
-require_relative "source2epub/source2epub"
-require_relative "source2epub/exporter"
-require_relative "source2epub/cli"
+# require 'rubygems'
+# require 'bundler/setup'
+
+# Community's gems
+require "git"
+require "thor"
+require "uri"
+require "eeepub"
+# My own gems
+require "agile_utils"
+require "code_lister"
+require "vim_printer"
+
+require "config/source2epub"
+require "source2epub/configuration"
+require "source2epub/version"
+require "source2epub/source2epub"
+require "source2epub/exporter"
+require "source2epub/cli"
 include Source2Epub